# web/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tamandua",
    "version": "0.9"
  },
  "paths": {
...
    "/api/runs/page": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Runs Page",
        "description": "`type` accepts several kinds separated by commas. Paging values out of range are clamped.",
        "operationId": "runs_page_api_runs_page_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "25",
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0",
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Type"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "asset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Asset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Run Detail",
        "operationId": "run_detail_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Detail Api Runs  Run Id  Get"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/{artifact}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Run Artifact",
        "description": "`title` names a compliance profile report (report-<profile>.md|pdf).",
        "operationId": "run_artifact_api_runs__run_id___artifact__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "artifact",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact"
            }
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Title"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The file: Markdown or PDF report, SARIF, CycloneDX SBOM, OpenVEX or Jira tickets (JSON)",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/markdown": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "maxLength": 50000000
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "maxLength": 50000000
                }
              },
              "application/sarif+json": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "maxLength": 100000000
                }
              },
              "application/vnd.cyclonedx+json": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "maxLength": 100000000
                }
              }
            }
          },
          "404": {
            "description": "Run not found, or that format isn't available for it"
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RunPage": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "RunPage"
      },
...
      "Error": {
        "title": "Error",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "In the reader's language."
          }
        }
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tamandua_session",
        "description": "Session signed in to the panel."
      },
      "metrics": {
        "type": "http",
        "scheme": "bearer",
        "description": "TAMANDUA_METRICS_TOKEN (Prometheus scraper)."
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ]
}